requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
whitenoise==6.0.0
